refactor(connector): use AwsAuthProps instead of HashMap in config
